Add API for updating ticket details by the requester.
Updating ticket details can be done only by the requester of the ticket only till the ticket is in assigned state. The route for this previously was ticket/:id/edit but cancancan was confused between this route and default edit route, to resolve this, the update ticket route was changed to tickets/:id/update_ticket_progress and this API to tickets/:id.